Moved auto-hide logic at SearchBarV2 level.
