+++ src/components/search/SearchDropdown.vue
@@ -23,7 +23,7 @@
 <!-- --------------------------------------------------------------------------------------------------------------- -->
 
 <template>
-  <div style="position: relative" ref="root" data-cy="searchDropdown">
+  <div style="position: relative" data-cy="searchDropdown">
     <div v-if="searchController.visible.value" class="box" style="position: absolute; display: flex; flex-direction: column; gap: 1rem; width: 100%; top: 5px; left: 0; z-index: 10; border: 0.5px solid white; padding: 16px 12px;">
       <template v-for="(c,i) in searchController.candidates.value" :key="c.description">
         <button class="button-as-link h-is-property-text"
@@ -53,7 +53,7 @@
 
 <script setup lang="ts">
 
-import {onBeforeUnmount, onMounted, PropType, ref} from "vue";
+import {PropType} from "vue";
 import {SearchController} from "@/components/search/SearchController";
 import {SearchCandidate} from "@/components/search/SearchAgent";
 import router from "@/router";
@@ -70,21 +70,6 @@ const navigate = (c: SearchCandidate<unknown>) => {
   c.agent.willNavigate(c)
   router.push(c.route)
 }
-
-const root = ref<HTMLElement|null>(null)
-const isInside = (target: Node) => root.value !== null && root.value.contains(target)
-
-const onMouseDown = (ev: MouseEvent) => {
-  if (ev.target instanceof Node && !isInside(ev.target)) {
-    props.searchController.inputText.value = "" // Hides SearchDropdown
-  }
-}
-onMounted(() => {
-  document.addEventListener("mousedown", onMouseDown)
-})
-onBeforeUnmount(() => {
-  document.removeEventListener("mousedown", onMouseDown)
-})
 
 </script>
